fix: directory refactor
