+++ src/curve/field/scalar.rs
@@ -7,12 +7,13 @@ use crypto_bigint::{
     Encoding, NonZero, U448,
 };
 
-pub const R_2: &str = "049b9b60e3539257c1b195d97af32c4b88ea18590d66de235ee4d838ae17cf72a3c47c441a9cc14be4d070af2052bcb7f823b7293402a939";
+/// Order of the curve. Distinct from field prime.
+pub const R_448: &str = "3FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF7CCA23E9C44EDB49AED63690216CC2728DC58F552378C292AB5844F3";
 
 impl_modulus!(
     R,
     U448,
-    "3FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF7CCA23E9C44EDB49AED63690216CC2728DC58F552378C292AB5844F3"
+    R_448
 );
 
 #[derive(Debug, Clone, Copy)]
